+++ notifications/subscriptions/raw_client.go
@@ -6,11 +6,11 @@ import (
 	context "context"
 	http "net/http"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v2"
-	core "github.com/KardFinancial/kard-go-sdk/v2/core"
-	internal "github.com/KardFinancial/kard-go-sdk/v2/internal"
-	notifications "github.com/KardFinancial/kard-go-sdk/v2/notifications"
-	option "github.com/KardFinancial/kard-go-sdk/v2/option"
+	kard "github.com/KardFinancial/kard-go-sdk/v3"
+	core "github.com/KardFinancial/kard-go-sdk/v3/core"
+	internal "github.com/KardFinancial/kard-go-sdk/v3/internal"
+	notifications "github.com/KardFinancial/kard-go-sdk/v3/notifications"
+	option "github.com/KardFinancial/kard-go-sdk/v3/option"
 )
 
 type RawClient struct {