+++ notifications/subscriptions_test.go
@@ -4,7 +4,7 @@ package notifications
 
 import (
 	json "encoding/json"
-	kardgosdk "github.com/KardFinancial/kard-go-sdk/v2"
+	kardgosdk "github.com/KardFinancial/kard-go-sdk/v3"
 	assert "github.com/stretchr/testify/assert"
 	require "github.com/stretchr/testify/require"
 	testing "testing"