+++ organizations/children/client.go
@@ -6,11 +6,11 @@ import (
 	context "context"
 	os "os"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v2"
-	core "github.com/KardFinancial/kard-go-sdk/v2/core"
-	internal "github.com/KardFinancial/kard-go-sdk/v2/internal"
-	option "github.com/KardFinancial/kard-go-sdk/v2/option"
-	organizations "github.com/KardFinancial/kard-go-sdk/v2/organizations"
+	kard "github.com/KardFinancial/kard-go-sdk/v3"
+	core "github.com/KardFinancial/kard-go-sdk/v3/core"
+	internal "github.com/KardFinancial/kard-go-sdk/v3/internal"
+	option "github.com/KardFinancial/kard-go-sdk/v3/option"
+	organizations "github.com/KardFinancial/kard-go-sdk/v3/organizations"
 )
 
 type Client struct {