+++ organizations/client/client.go
@@ -6,13 +6,13 @@ import (
 	context "context"
 	os "os"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v2"
-	core "github.com/KardFinancial/kard-go-sdk/v2/core"
-	internal "github.com/KardFinancial/kard-go-sdk/v2/internal"
-	option "github.com/KardFinancial/kard-go-sdk/v2/option"
-	children "github.com/KardFinancial/kard-go-sdk/v2/organizations/children"
-	contentstrategies "github.com/KardFinancial/kard-go-sdk/v2/organizations/contentstrategies"
-	placements "github.com/KardFinancial/kard-go-sdk/v2/organizations/placements"
+	kard "github.com/KardFinancial/kard-go-sdk/v3"
+	core "github.com/KardFinancial/kard-go-sdk/v3/core"
+	internal "github.com/KardFinancial/kard-go-sdk/v3/internal"
+	option "github.com/KardFinancial/kard-go-sdk/v3/option"
+	children "github.com/KardFinancial/kard-go-sdk/v3/organizations/children"
+	contentstrategies "github.com/KardFinancial/kard-go-sdk/v3/organizations/contentstrategies"
+	placements "github.com/KardFinancial/kard-go-sdk/v3/organizations/placements"
 )
 
 type Client struct {